Version: 3.0, Date: 2024-05-01, Build: 5 -- Added new admin terms management page and enhanced user interface
A new management page for terms has been created which includes the ability to add new terms and comes with a layout for updating and deleting terms. Additionally, user information has been added to the session upon successful login and this information is now displayed in the header of all admin pages. The absolute path to the SQLite database has also been corrected. The interface of the dropdown user box has been optimized, including the logout function.