Version: 3.0, Date: 2024-05-06, Build: 22 -- Remove backend logger, add log file error handling in adminLogger.php
The backend logger.php file has been removed. The adminLogger.php template was updated to include error handling for situations when the log file cannot be accessed or does not exist. This update displays an appropriate error message to the admin. Additionally, the textarea CSS styles in adminLogger.css have been adjusted.