Added course competency editing feature for professor
Implemented new functionality allowing professors to edit the competence of courses. This included creation of a new form in the professor's page, addition of a CSS class for styling the action button, and a new page for handling the editing process of course competency. Required CSS for the new page was also created.